Fixing tabId import location
